Remove more warnings from the codebase

+++ modules/telemetry/src/general/timetelemetry.cpp
@@ -76,7 +76,7 @@ TimeTelemetry::TimeTelemetry(const std::string& ip, int port)
     , _precisionProperties(TimeTelemetry::PrecisionProperties(PrecisionInfo))
 {
     for (size_t i = 0; i < TimeUnitNames.size(); ++i) {
-        _timeUnitOption.addOption(i, TimeUnitNames[i].singular.data());
+        _timeUnitOption.addOption(static_cast<int>(i), TimeUnitNames[i].singular.data());
     }
 
     _timeUnitOption.setValue(static_cast<int>(TimeUnit::Day));
